Updates to simulation script and bash script addition for simulating multiple UAVs with Vagrant on mac
Unfortunately for testing purposes, it's a pain to multithread multiple dronekit-sitl instances and test the dronekit scripts that will go on each drone. Therefore I had to natively install SITL (software in the loop), which requires Vagrant for setting up a Linux VM if you're using Mac. If you're using Windows or Linux, this can be installed natively on the OS. From such, there are straightforward arguments in the native SITL build which allow for the creation of multiple instances, which are used by the new DroneKit_simulator script.

Because of this work, the commits here are mostly intermediate steps (related to Mac) before I move forward on building the entire simulation environment on any system. The bash script "gather_scripts" is used on a Mac to create multiple instances of SITL from which "DroneKit_simulator" then attaches to. This requires a Vagrant install of SITL on Mac, which is found in the dronekit/ardupilot resources. If you wish to run this on Windows or Linux, you'll have to first install SITL natively then setup your bash script that will create multiple SITL instances then ensure the main script is connected to the correct forwarding ports displayed by these SITL instances.

I'm planning to cleanup this process after we've decided on the architecture for entire system of drones.